refactor(LottoGeneratorService): 클래스명 변경 및 패키지 이전
로또 생성 기능을 서비스 객체로 할당함에 따라 클래스명을 변경하고 패키지를 이전합니다.
